Extend group operations to support bootstrap
Sometimes participants wanting to form a group
have limited knowledge of their co-participants.
For example, two procs may know that they "lead"
their respective subgroups, but may not know the
members of the other "leaders" subgroup. In this
case, we need to let the two procs that know about
their subgroups tell us what they know, and let
the other procs simply join the collection.

All we really require is that everyone know the
groupID for the group they are trying to form, and
that each leader know how many leaders are going to
participate. We can use events under-the-covers to
notify non-leaders of the group formation.

This is a collaborative change with PRRTE.
