+++ .gitignore
@@ -100,6 +100,7 @@ examples/abi_with_init
 examples/group_lcl_cid
 examples/nodeinfo
 examples/pset
+examples/group_bootstrap
 
 include/pmix_version.h
 include/pmix_rename.h
