+++ docs/how-things-work/index.rst
@@ -12,3 +12,4 @@ find information on that subject here.
    session_dirs.rst
    distances.rst
    schedulers/index.rst
+   sets_groups/index.rst
