+++ docs/how-things-work/sets_groups/index.rst
@@ -1,0 +1,12 @@
+
+Process Sets and Groups
+=======================
+
+This section describes PMIx support for process
+sets and groups.
+
+.. toctree::
+   :maxdepth: 2
+
+   overview.rst
+   psets.rst
