+++ src/server/pmix_server_ops.c
@@ -357,15 +357,19 @@ pmix_status_t pmix_server_commit(pmix_peer_t *peer, pmix_buffer_t *buf)
  *         regardless of location
  * nprocs - the number of procs in the array
  */
-static pmix_server_trkr_t *get_tracker(char *id, pmix_proc_t *procs,
-                                       size_t nprocs, pmix_cmd_t type)
+pmix_server_trkr_t *pmix_server_get_tracker(char *id, pmix_proc_t *procs,
+                                            size_t nprocs, pmix_cmd_t type)
 {
     pmix_server_trkr_t *trk;
-    size_t i, j;
+    size_t i, j, sz;
     size_t matches;
+    pmix_proclist_t *p;
+    pmix_proc_t *ptr;
+    pmix_list_t cache;
+    bool found;
 
     pmix_output_verbose(5, pmix_server_globals.fence_output,
-                        "get_tracker called with %d procs",
+                        "pmix_server_get_tracker called with %d procs",
                         (int) nprocs);
 
     /* bozo check - should never happen outside of programmer error */
@@ -387,6 +391,40 @@ static pmix_server_trkr_t *get_tracker(char *id, pmix_proc_t *procs,
          */
         if (NULL != id) {
             if (NULL != trk->id && 0 == strcmp(id, trk->id)) {
+                PMIX_CONSTRUCT(&cache, pmix_list_t);
+                /* update tracked procs to include the given ones,
+                 * filtered to only keep unique entries */
+                for (i=0; i < nprocs; i++) {
+                    found = false;
+                    for (j=0; j < trk->npcs; j++) {
+                        if (PMIX_CHECK_PROCID(&procs[i], &trk->pcs[j])) {
+                            // match - can ignore it
+                            found = true;
+                            break;
+                        }
+                    }
+                    if (!found) {
+                        // new proc, so we need to add it
+                        p = PMIX_NEW(pmix_proclist_t);
+                        memcpy(&p->proc, &procs[i], sizeof(pmix_proc_t));
+                        pmix_list_append(&cache, &p->super);
+                    }
+                }
+                i = pmix_list_get_size(&cache);
+                if (0 < i) {
+                    sz = trk->npcs + i;
+                    PMIX_PROC_CREATE(ptr, sz);
+                    memcpy(ptr, trk->pcs, trk->npcs * sizeof(pmix_proc_t));
+                    j = trk->npcs;
+                    PMIX_LIST_FOREACH(p, &cache, pmix_proclist_t) {
+                        memcpy(&ptr[j], &p->proc, sizeof(pmix_proc_t));
+                        ++j;
+                    }
+                    PMIX_PROC_FREE(trk->pcs, trk->npcs);
+                    trk->pcs = ptr;
+                    trk->npcs = sz;
+                }
+                PMIX_LIST_DESTRUCT(&cache);
                 return trk;
             }
         } else {
@@ -433,8 +471,8 @@ static pmix_server_trkr_t *get_tracker(char *id, pmix_proc_t *procs,
  *         regardless of location
  * nprocs - the number of procs in the array
  */
-static pmix_server_trkr_t *new_tracker(char *id, pmix_proc_t *procs,
-                                       size_t nprocs, pmix_cmd_t type)
+pmix_server_trkr_t *pmix_server_new_tracker(char *id, pmix_proc_t *procs,
+                                            size_t nprocs, pmix_cmd_t type)
 {
     pmix_server_trkr_t *trk;
     size_t i;
@@ -445,7 +483,7 @@ static pmix_server_trkr_t *new_tracker(char *id, pmix_proc_t *procs,
     pmix_nspace_t first;
 
     pmix_output_verbose(5, pmix_server_globals.fence_output,
-                        "new_tracker called with %d procs",
+                        "pmix_server_new_tracker called with %d procs",
                         (int) nprocs);
 
     /* bozo check - should never happen outside of programmer error */
@@ -543,7 +581,7 @@ static pmix_server_trkr_t *new_tracker(char *id, pmix_proc_t *procs,
         if (0 == nptr->nlocalprocs) {
             /* the host has informed us that this nspace has no local procs */
             pmix_output_verbose(5, pmix_server_globals.fence_output,
-                                "new_tracker: nspace %s has no local procs", procs[i].nspace);
+                                "pmix_server_new_tracker: nspace %s has no local procs", procs[i].nspace);
             trk->local = false;
             continue;
         }
@@ -589,7 +627,7 @@ static pmix_server_trkr_t *new_tracker(char *id, pmix_proc_t *procs,
              * process it once all the procs are known */
             all_def = false;
             pmix_output_verbose(5, pmix_server_globals.fence_output,
-                                "new_tracker: all clients not registered nspace %s",
+                                "pmix_server_new_tracker: all clients not registered nspace %s",
                                 procs[i].nspace);
             continue;
         }
@@ -634,8 +672,8 @@ static void fence_timeout(int sd, short args, void *cbdata)
     PMIX_RELEASE(trk);
 }
 
-static pmix_status_t _collect_data(pmix_server_trkr_t *trk,
-                                   pmix_buffer_t *buf)
+pmix_status_t pmix_server_collect_data(pmix_server_trkr_t *trk,
+                                       pmix_buffer_t *buf)
 {
     pmix_buffer_t bucket, *pbkt = NULL;
     pmix_cb_t cb;
@@ -963,9 +1001,9 @@ pmix_status_t pmix_server_fence(pmix_server_caddy_t *cd, pmix_buffer_t *buf,
     }
 
     /* find/create the local tracker for this operation */
-    if (NULL == (trk = get_tracker(NULL, procs, nprocs, PMIX_FENCENB_CMD))) {
+    if (NULL == (trk = pmix_server_get_tracker(NULL, procs, nprocs, PMIX_FENCENB_CMD))) {
         /* If no tracker was found - create and initialize it once */
-        if (NULL == (trk = new_tracker(NULL, procs, nprocs, PMIX_FENCENB_CMD))) {
+        if (NULL == (trk = pmix_server_new_tracker(NULL, procs, nprocs, PMIX_FENCENB_CMD))) {
             /* only if a bozo error occurs */
             PMIX_ERROR_LOG(PMIX_ERROR);
             /* DO NOT HANG */
@@ -1080,7 +1118,7 @@ pmix_status_t pmix_server_fence(pmix_server_caddy_t *cd, pmix_buffer_t *buf,
          * or global distribution */
 
         PMIX_CONSTRUCT(&bucket, pmix_buffer_t);
-        if (PMIX_SUCCESS != (rc = _collect_data(trk, &bucket))) {
+        if (PMIX_SUCCESS != (rc = pmix_server_collect_data(trk, &bucket))) {
             PMIX_ERROR_LOG(rc);
             PMIX_DESTRUCT(&bucket);
             /* clear the caddy from this tracker so it can be
@@ -1765,9 +1803,9 @@ pmix_status_t pmix_server_disconnect(pmix_server_caddy_t *cd, pmix_buffer_t *buf
     }
 
     /* find/create the local tracker for this operation */
-    if (NULL == (trk = get_tracker(NULL, procs, nprocs, PMIX_DISCONNECTNB_CMD))) {
+    if (NULL == (trk = pmix_server_get_tracker(NULL, procs, nprocs, PMIX_DISCONNECTNB_CMD))) {
         /* we don't have this tracker yet, so get a new one */
-        if (NULL == (trk = new_tracker(NULL, procs, nprocs, PMIX_DISCONNECTNB_CMD))) {
+        if (NULL == (trk = pmix_server_new_tracker(NULL, procs, nprocs, PMIX_DISCONNECTNB_CMD))) {
             /* only if a bozo error occurs */
             PMIX_ERROR_LOG(PMIX_ERROR);
             rc = PMIX_ERROR;
@@ -1944,9 +1982,9 @@ pmix_status_t pmix_server_connect(pmix_server_caddy_t *cd, pmix_buffer_t *buf,
     }
 
     /* find/create the local tracker for this operation */
-    if (NULL == (trk = get_tracker(NULL, procs, nprocs, PMIX_CONNECTNB_CMD))) {
+    if (NULL == (trk = pmix_server_get_tracker(NULL, procs, nprocs, PMIX_CONNECTNB_CMD))) {
         /* we don't have this tracker yet, so get a new one */
-        if (NULL == (trk = new_tracker(NULL, procs, nprocs, PMIX_CONNECTNB_CMD))) {
+        if (NULL == (trk = pmix_server_new_tracker(NULL, procs, nprocs, PMIX_CONNECTNB_CMD))) {
             /* only if a bozo error occurs */
             PMIX_ERROR_LOG(PMIX_ERROR);
             /* DO NOT HANG */
@@ -3716,1135 +3754,6 @@ error:
     return rc;
 }
 
-static void _grpcbfunc(int sd, short args, void *cbdata)
-{
-    pmix_shift_caddy_t *scd = (pmix_shift_caddy_t *) cbdata;
-    pmix_server_trkr_t *trk = scd->tracker;
-    pmix_server_caddy_t *cd;
-    pmix_buffer_t *reply, xfer, dblob, rankblob;
-    pmix_status_t ret;
-    size_t n, ctxid = SIZE_MAX, ngrpinfo, nmembers = 0;
-    pmix_proc_t *members = NULL;
-    pmix_byte_object_t *bo = NULL, pbo;
-    pmix_nspace_caddy_t *nptr;
-    pmix_list_t nslist;
-    bool found;
-    bool ctxid_given = false;
-    pmix_regattr_input_t *p;
-    uint32_t index, endptidx, infoidx;
-    int32_t cnt;
-    pmix_proc_t procid;
-    pmix_info_t *grpinfo, *iptr;
-    pmix_kval_t kp;
-    pmix_value_t val;
-    pmix_data_array_t darray;
-
-    PMIX_ACQUIRE_OBJECT(scd);
-    PMIX_HIDE_UNUSED_PARAMS(sd, args);
-
-    if (NULL == trk) {
-        /* give them a release if they want it - this should
-         * never happen, but protect against the possibility */
-        if (NULL != scd->cbfunc.relfn) {
-            scd->cbfunc.relfn(scd->cbdata);
-        }
-        PMIX_RELEASE(scd);
-        return;
-    }
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "server:grpcbfunc processing WITH %d CALLBACKS",
-                        (int) pmix_list_get_size(&trk->local_cbs));
-
-    /* the tracker's "hybrid" field is used to indicate construct
-     * vs destruct - true is destruct */
-    if (trk->hybrid) {
-        goto release;
-    }
-
-    /* see if this group was assigned a context ID or collected data */
-    for (n = 0; n < scd->ninfo; n++) {
-        if (PMIX_CHECK_KEY(&scd->info[n], PMIX_GROUP_CONTEXT_ID)) {
-            PMIX_VALUE_GET_NUMBER(ret, &scd->info[n].value, ctxid, size_t);
-            if (PMIX_SUCCESS != ret) {
-                PMIX_ERROR_LOG(ret);
-            } else {
-                ctxid_given = true;
-            }
-        } else if (PMIX_CHECK_KEY(&scd->info[n], PMIX_GROUP_ENDPT_DATA)) {
-            bo = &scd->info[n].value.data.bo;
-        } else if (PMIX_CHECK_KEY(&scd->info[n], PMIX_GROUP_MEMBERSHIP)) {
-            members = (pmix_proc_t*)scd->info[n].value.data.darray->array;
-            nmembers = scd->info[n].value.data.darray->size;
-        }
-    }
-
-    /* if data was returned, then we need to have the modex cbfunc
-     * store it for us before releasing the group members */
-    if (NULL != bo) {
-        /* get the indices of the types of data */
-        p = pmix_hash_lookup_key(UINT32_MAX, PMIX_GROUP_ENDPT_DATA, NULL);
-        endptidx = p->index;
-        p = pmix_hash_lookup_key(UINT32_MAX, PMIX_GROUP_INFO, NULL);
-        infoidx = p->index;
-
-        PMIX_CONSTRUCT(&xfer, pmix_buffer_t);
-        PMIX_LOAD_BUFFER(pmix_globals.mypeer, &xfer, bo->bytes, bo->size);
-        PMIX_CONSTRUCT(&nslist, pmix_list_t);
-        /* Collect the nptr list with uniq GDS components of all local
-         * participants. It does not allow multiple storing to the
-         * same GDS if participants have mutual GDS. */
-        PMIX_LIST_FOREACH (cd, &trk->local_cbs, pmix_server_caddy_t) {
-            // see if we already have this nspace
-            found = false;
-            PMIX_LIST_FOREACH (nptr, &nslist, pmix_nspace_caddy_t) {
-                if (0 == strcmp(nptr->ns->compat.gds->name, cd->peer->nptr->compat.gds->name)) {
-                    found = true;
-                    break;
-                }
-            }
-            if (!found) {
-                // add it
-                nptr = PMIX_NEW(pmix_nspace_caddy_t);
-                PMIX_RETAIN(cd->peer->nptr);
-                nptr->ns = cd->peer->nptr;
-                pmix_list_append(&nslist, &nptr->super);
-            }
-        }
-
-        /* unpack each byte object - we get one byte object containing the
-         * complete contribution from each participating server. Each byte
-         * object contains up to two byte objects - one for the endpt data
-         * from the local participants, and the other potentially containing
-         * any group info published by those participants */
-        /* unpack the index indicating the type of data in the object */
-        cnt = 1;
-        PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &xfer, &index, &cnt, PMIX_UINT32);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_ERROR_LOG(ret);
-            PMIX_DESTRUCT(&xfer);
-            goto release;
-        }
-        while (PMIX_SUCCESS == ret) {
-            /* unpack the data blob for this object */
-            cnt = 1;
-            PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &xfer, &pbo, &cnt, PMIX_BYTE_OBJECT);
-            if (PMIX_SUCCESS != ret) {
-                PMIX_ERROR_LOG(ret);
-                PMIX_DESTRUCT(&xfer);
-                goto release;
-            }
-            PMIX_CONSTRUCT(&dblob, pmix_buffer_t);
-            PMIX_LOAD_BUFFER(pmix_globals.mypeer, &dblob, pbo.bytes, pbo.size);
-            PMIX_BYTE_OBJECT_DESTRUCT(&pbo);
-
-            if (index == endptidx) {
-                PMIX_LIST_FOREACH (nptr, &nslist, pmix_nspace_caddy_t) {
-                    PMIX_GDS_STORE_MODEX(ret, nptr->ns, &dblob, trk);
-                    if (PMIX_SUCCESS != ret) {
-                        PMIX_ERROR_LOG(ret);
-                        break;
-                    }
-                }
-            } else if (index == infoidx) {
-                /* this is packed differently, at least for now, so we have
-                 * to unpack it and process it directly */
-                if (ctxid_given) {
-                    /* the blob consists of a set of byte objects, each containing the ID
-                     * of the contributing proc followed by the pmix_info_t they
-                     * provided */
-                    ret = PMIX_SUCCESS;
-                    while (PMIX_SUCCESS == ret) {
-                        cnt = 1;
-                        PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &dblob, &pbo, &cnt, PMIX_BYTE_OBJECT);
-                        if (PMIX_ERR_UNPACK_READ_PAST_END_OF_BUFFER == ret) {
-                            /* not an error - we are simply done */
-                            break;
-                        }
-                        if (PMIX_SUCCESS != ret) {
-                            PMIX_ERROR_LOG(ret);
-                            PMIX_DESTRUCT(&xfer);
-                            PMIX_DESTRUCT(&dblob);
-                            goto release;
-                        }
-                        PMIX_CONSTRUCT(&rankblob, pmix_buffer_t);
-                        PMIX_LOAD_BUFFER(pmix_globals.mypeer, &rankblob, pbo.bytes, pbo.size);
-                        PMIX_BYTE_OBJECT_DESTRUCT(&pbo);
-                        cnt = 1;
-                        PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &rankblob, &procid, &cnt, PMIX_PROC);
-                        if (PMIX_SUCCESS != ret) {
-                            PMIX_ERROR_LOG(ret);
-                            PMIX_DESTRUCT(&xfer);
-                            PMIX_DESTRUCT(&dblob);
-                            PMIX_DESTRUCT(&rankblob);
-                            goto release;
-                        }
-                        cnt = 1;
-                        PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &rankblob, &ngrpinfo, &cnt, PMIX_SIZE);
-                        if (PMIX_SUCCESS != ret) {
-                            PMIX_ERROR_LOG(ret);
-                            PMIX_DESTRUCT(&xfer);
-                            PMIX_DESTRUCT(&dblob);
-                            PMIX_DESTRUCT(&rankblob);
-                            goto release;
-                        }
-                        PMIX_INFO_CREATE(grpinfo, ngrpinfo);
-                        cnt = ngrpinfo;
-                        PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &rankblob, grpinfo, &cnt, PMIX_INFO);
-                        if (PMIX_SUCCESS != ret) {
-                            PMIX_ERROR_LOG(ret);
-                            PMIX_DESTRUCT(&xfer);
-                            PMIX_DESTRUCT(&dblob);
-                            PMIX_DESTRUCT(&rankblob);
-                            PMIX_INFO_FREE(grpinfo, ngrpinfo);
-                            goto release;
-                        }
-                        /* reconstruct each value as a qualified one basd
-                         * on the ctxid */
-                        PMIX_CONSTRUCT(&kp, pmix_kval_t);
-                        kp.value = &val;
-                        kp.key = PMIX_QUALIFIED_VALUE;
-                        val.type = PMIX_DATA_ARRAY;
-                        for (n=0; n < ngrpinfo; n++) {
-                            PMIX_DATA_ARRAY_CONSTRUCT(&darray, 2, PMIX_INFO);
-                            iptr = (pmix_info_t*)darray.array;
-                            /* the primary value is in the first position */
-                            PMIX_INFO_XFER(&iptr[0], &grpinfo[n]);
-                            /* add the context ID qualifier */
-                            PMIX_INFO_LOAD(&iptr[1], PMIX_GROUP_CONTEXT_ID, &ctxid, PMIX_SIZE);
-                            PMIX_INFO_SET_QUALIFIER(&iptr[1]);
-                            /* add it to the kval */
-                            val.data.darray = &darray;
-                            /* store it */
-                            PMIX_GDS_STORE_KV(ret, pmix_globals.mypeer, &procid, PMIX_GLOBAL, &kp);
-                            PMIX_DATA_ARRAY_DESTRUCT(&darray);
-                            if (PMIX_SUCCESS != ret) {
-                                PMIX_ERROR_LOG(ret);
-                                PMIX_DESTRUCT(&xfer);
-                                PMIX_DESTRUCT(&dblob);
-                                PMIX_DESTRUCT(&rankblob);
-                                PMIX_INFO_FREE(grpinfo, ngrpinfo);
-                                goto release;
-                            }
-                        }
-                        PMIX_INFO_FREE(grpinfo, ngrpinfo);
-                        PMIX_DESTRUCT(&rankblob);
-                        ret = PMIX_SUCCESS;
-                    }
-                }
-            }
-            PMIX_DESTRUCT(&dblob);
-            /* get the next blob */
-            cnt = 1;
-            PMIX_BFROPS_UNPACK(ret, pmix_globals.mypeer, &xfer, &index, &cnt, PMIX_UINT32);
-            if (PMIX_SUCCESS != ret && PMIX_ERR_UNPACK_READ_PAST_END_OF_BUFFER != ret) {
-                PMIX_ERROR_LOG(ret);
-                break;
-            }
-        }
-        PMIX_DESTRUCT(&xfer);
-    }
-
-release:
-    /* loop across all procs in the tracker, sending them the reply */
-    PMIX_LIST_FOREACH (cd, &trk->local_cbs, pmix_server_caddy_t) {
-        reply = PMIX_NEW(pmix_buffer_t);
-        if (NULL == reply) {
-            break;
-        }
-        /* setup the reply, starting with the returned status */
-        PMIX_BFROPS_PACK(ret, cd->peer, reply, &scd->status, 1, PMIX_STATUS);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_ERROR_LOG(ret);
-            PMIX_RELEASE(reply);
-            break;
-        }
-        if (!trk->hybrid) {
-            /* add the final membership */
-            PMIX_BFROPS_PACK(ret, cd->peer, reply, &nmembers, 1, PMIX_SIZE);
-            if (PMIX_SUCCESS != ret) {
-                PMIX_ERROR_LOG(ret);
-                PMIX_RELEASE(reply);
-                break;
-            }
-            if (0 < nmembers) {
-                PMIX_BFROPS_PACK(ret, cd->peer, reply, members, nmembers, PMIX_PROC);
-                if (PMIX_SUCCESS != ret) {
-                    PMIX_ERROR_LOG(ret);
-                    PMIX_RELEASE(reply);
-                    break;
-                }
-            }
-            /* if a ctxid was provided, pass it along */
-            PMIX_BFROPS_PACK(ret, cd->peer, reply, &ctxid, 1, PMIX_SIZE);
-            if (PMIX_SUCCESS != ret) {
-                PMIX_ERROR_LOG(ret);
-                PMIX_RELEASE(reply);
-                break;
-            }
-        }
-        pmix_output_verbose(2, pmix_server_globals.group_output,
-                            "server:grp_cbfunc reply being sent to %s:%u",
-                            cd->peer->info->pname.nspace, cd->peer->info->pname.rank);
-        PMIX_SERVER_QUEUE_REPLY(ret, cd->peer, cd->hdr.tag, reply);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_RELEASE(reply);
-        }
-    }
-
-    /* remove the tracker from the list */
-    pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-    PMIX_RELEASE(trk);
-
-    /* we are done */
-    if (NULL != scd->cbfunc.relfn) {
-        scd->cbfunc.relfn(scd->cbdata);
-    }
-    PMIX_RELEASE(scd);
-}
-
-static void grpcbfunc(pmix_status_t status,
-                      pmix_info_t *info, size_t ninfo, void *cbdata,
-                      pmix_release_cbfunc_t relfn, void *relcbd)
-{
-    pmix_server_trkr_t *tracker = (pmix_server_trkr_t *) cbdata;
-    pmix_shift_caddy_t *scd;
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "server:grpcbfunc called with %d info", (int) ninfo);
-
-    if (NULL == tracker) {
-        /* nothing to do - but be sure to give them
-         * a release if they want it */
-        if (NULL != relfn) {
-            relfn(relcbd);
-        }
-        return;
-    }
-    /* need to thread-shift this callback as it accesses global data */
-    scd = PMIX_NEW(pmix_shift_caddy_t);
-    if (NULL == scd) {
-        /* nothing we can do */
-        if (NULL != relfn) {
-            relfn(cbdata);
-        }
-        return;
-    }
-    scd->status = status;
-    scd->info = info;
-    scd->ninfo = ninfo;
-    scd->tracker = tracker;
-    scd->cbfunc.relfn = relfn;
-    scd->cbdata = relcbd;
-    PMIX_THREADSHIFT(scd, _grpcbfunc);
-}
-
-static void grp_timeout(int sd, short args, void *cbdata)
-{
-    pmix_server_trkr_t *trk = (pmix_server_trkr_t *) cbdata;
-    pmix_server_caddy_t *cd;
-    pmix_buffer_t *reply;
-    pmix_status_t ret, rc = PMIX_ERR_TIMEOUT;
-    PMIX_HIDE_UNUSED_PARAMS(sd, args);
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "ALERT: grp construct timeout fired");
-
-    /* loop across all procs in the tracker, alerting
-     * them to the failure */
-    PMIX_LIST_FOREACH (cd, &trk->local_cbs, pmix_server_caddy_t) {
-        reply = PMIX_NEW(pmix_buffer_t);
-        if (NULL == reply) {
-            break;
-        }
-        /* setup the reply, starting with the returned status */
-        PMIX_BFROPS_PACK(ret, cd->peer, reply, &rc, 1, PMIX_STATUS);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_ERROR_LOG(ret);
-            PMIX_RELEASE(reply);
-            break;
-        }
-        pmix_output_verbose(2, pmix_server_globals.group_output,
-                            "server:grp_cbfunc TIMEOUT being sent to %s:%u",
-                            cd->peer->info->pname.nspace, cd->peer->info->pname.rank);
-        PMIX_SERVER_QUEUE_REPLY(ret, cd->peer, cd->hdr.tag, reply);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_RELEASE(reply);
-        }
-    }
-
-    trk->event_active = false;
-    /* record this group as failed */
-    PMIx_Argv_append_nosize(&pmix_server_globals.failedgrps, trk->id);
-    /* remove the tracker from the list */
-    pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-    PMIX_RELEASE(trk);
-}
-
-/* we are being called from the PMIx server's switchyard function,
- * which means we are in an event and can access global data */
-pmix_status_t pmix_server_grpconstruct(pmix_server_caddy_t *cd, pmix_buffer_t *buf)
-{
-    pmix_peer_t *peer = (pmix_peer_t *) cd->peer;
-    pmix_peer_t *pr;
-    int32_t cnt, m;
-    pmix_status_t rc;
-    char *grpid;
-    pmix_proc_t *procs;
-    pmix_info_t *info = NULL, *iptr = NULL, *grpinfoptr = NULL;
-    size_t n, ninfo, ninf, niptr, nprocs, n2, ngrpinfo = 0;
-    pmix_server_trkr_t *trk;
-    bool need_cxtid = false;
-    bool match, force_local = false;
-    bool embed_barrier = false;
-    bool barrier_directive_included = false;
-    bool locally_complete = false;
-    pmix_buffer_t bucket, bkt;
-    pmix_byte_object_t bo;
-    pmix_grpinfo_t *g = NULL;
-    pmix_regattr_input_t *p;
-    struct timeval tv = {0, 0};
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "recvd grpconstruct cmd from %s",
-                        PMIX_PEER_PRINT(cd->peer));
-
-    /* unpack the group ID */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &grpid, &cnt, PMIX_STRING);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-
-    /* is this a failed group? */
-    if (NULL != pmix_server_globals.failedgrps) {
-        for (m=0; NULL != pmix_server_globals.failedgrps[m]; m++) {
-            if (0 == strcmp(grpid, pmix_server_globals.failedgrps[m])) {
-                /* yes - reject it */
-                free(grpid);
-                return PMIX_ERR_TIMEOUT;
-            }
-        }
-    }
-
-    /* unpack the number of procs */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &nprocs, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    if (0 == nprocs) {
-        PMIX_ERROR_LOG(PMIX_ERR_BAD_PARAM);
-        return PMIX_ERR_BAD_PARAM;
-    }
-    PMIX_PROC_CREATE(procs, nprocs);
-    if (NULL == procs) {
-        rc = PMIX_ERR_NOMEM;
-        goto error;
-    }
-    cnt = nprocs;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, procs, &cnt, PMIX_PROC);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_PROC_FREE(procs, nprocs);
-        goto error;
-    }
-    /* sort the procs */
-    qsort(procs, nprocs, sizeof(pmix_proc_t), pmix_util_compare_proc);
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ninf, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    ninfo = ninf + 2;
-    PMIX_INFO_CREATE(info, ninfo);
-    /* store default response */
-    rc = PMIX_SUCCESS;
-    PMIX_INFO_LOAD(&info[ninf], PMIX_SORTED_PROC_ARRAY, NULL, PMIX_BOOL);
-    PMIX_INFO_LOAD(&info[ninf+1], PMIX_LOCAL_COLLECTIVE_STATUS, &rc, PMIX_STATUS);
-    if (0 < ninf) {
-        cnt = ninf;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto error;
-        }
-    }
-
-    /* check directives */
-    for (n = 0; n < ninfo; n++) {
-        if (PMIX_CHECK_KEY(&info[n], PMIX_GROUP_ASSIGN_CONTEXT_ID)) {
-            need_cxtid = PMIX_INFO_TRUE(&info[n]);
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_GROUP_LOCAL_ONLY)) {
-            force_local = PMIX_INFO_TRUE(&info[n]);
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_EMBED_BARRIER)) {
-            embed_barrier = PMIX_INFO_TRUE(&info[n]);
-            barrier_directive_included = true;
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_TIMEOUT)) {
-            PMIX_VALUE_GET_NUMBER(rc, &info[n].value, tv.tv_sec, uint32_t);
-            if (PMIX_SUCCESS != rc) {
-                return rc;
-            }
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_GROUP_INFO)) {
-            grpinfoptr = (pmix_info_t*)info[n].value.data.darray->array;
-            ngrpinfo = info[n].value.data.darray->size;
-            g = PMIX_NEW(pmix_grpinfo_t);
-            PMIX_LOAD_NSPACE(g->proc.nspace, peer->nptr->nspace);
-            g->proc.rank = peer->info->pname.rank;
-            PMIX_CONSTRUCT(&bucket, pmix_buffer_t);
-            PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &g->proc, 1, PMIX_PROC);
-            if (PMIX_SUCCESS != rc) {
-                PMIX_ERROR_LOG(rc);
-                PMIX_RELEASE(g);
-                PMIX_DESTRUCT(&bucket);
-                goto error;
-            }
-            PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &ngrpinfo, 1, PMIX_SIZE);
-            if (PMIX_SUCCESS != rc) {
-                PMIX_ERROR_LOG(rc);
-                PMIX_RELEASE(g);
-                PMIX_DESTRUCT(&bucket);
-                goto error;
-            }
-            PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, grpinfoptr, ngrpinfo, PMIX_INFO);
-            if (PMIX_SUCCESS != rc) {
-                PMIX_ERROR_LOG(rc);
-                PMIX_RELEASE(g);
-                PMIX_DESTRUCT(&bucket);
-                goto error;
-            }
-            PMIX_UNLOAD_BUFFER(&bucket, g->blob.bytes, g->blob.size);
-            PMIX_DESTRUCT(&bucket);
-        }
-    }
-
-    /* find/create the local tracker for this operation */
-    trk = get_tracker(grpid, procs, nprocs, PMIX_GROUP_CONSTRUCT_CMD);
-    if (NULL == trk) {
-        /* If no tracker was found - create and initialize it once */
-        trk = new_tracker(grpid, procs, nprocs, PMIX_GROUP_CONSTRUCT_CMD);
-        if (NULL == trk) {
-            /* only if a bozo error occurs */
-            PMIX_ERROR_LOG(PMIX_ERROR);
-            rc = PMIX_ERROR;
-            goto error;
-        }
-        /* group members must have access to all endpoint info
-         * upon completion of the construct operation */
-        trk->collect_type = PMIX_COLLECT_YES;
-        /* mark as being a construct operation */
-        trk->hybrid = false;
-        /* it is possible that different participants will
-         * provide different attributes, so collect the
-         * aggregate of them */
-        if (NULL == trk->info) {
-            trk->info = info;
-            trk->ninfo = ninfo;
-        } else {
-            niptr = trk->ninfo + ninfo;
-            PMIX_INFO_CREATE(iptr, niptr);
-            for (n=0; n < trk->ninfo; n++) {
-                PMIX_INFO_XFER(&iptr[n], &trk->info[n]);
-            }
-            for (n=0; n < ninfo; n++) {
-                PMIX_INFO_XFER(&iptr[n+trk->ninfo], &info[n]);
-            }
-            PMIX_INFO_FREE(trk->info, trk->ninfo);
-            trk->info = iptr;
-            trk->ninfo = niptr;
-            /* cleanup */
-            PMIX_INFO_FREE(info, ninfo);
-            info = NULL;
-        }
-        if (NULL != g) {
-            pmix_list_append(&trk->grpinfo, &g->super);
-        }
-        /* see if this constructor only references local processes and isn't
-         * requesting a context ID - if both conditions are met, then we
-         * can just locally process the request without bothering the host.
-         * This is meant to provide an optimized path for a fairly common
-         * operation */
-        if (force_local) {
-            trk->local = true;
-        } else if (need_cxtid) {
-            trk->local = false;
-        } else {
-            trk->local = true;
-            for (n = 0; n < nprocs; n++) {
-                /* if this entry references the local procs, then
-                 * we can skip it */
-                if (PMIX_RANK_LOCAL_PEERS == procs[n].rank ||
-                    PMIX_RANK_LOCAL_NODE == procs[n].rank) {
-                    continue;
-                }
-                /* see if it references a specific local proc */
-                match = false;
-                for (m = 0; m < pmix_server_globals.clients.size; m++) {
-                    pr = (pmix_peer_t *) pmix_pointer_array_get_item(&pmix_server_globals.clients, m);
-                    if (NULL == pr) {
-                        continue;
-                    }
-                    if (PMIX_CHECK_NAMES(&procs[n], &pr->info->pname)) {
-                        match = true;
-                        break;
-                    }
-                }
-                if (!match) {
-                    /* this requires a non_local operation */
-                    trk->local = false;
-                    break;
-                }
-            }
-        }
-    } else {
-        /* it is possible that different participants will
-         * provide different attributes, so collect the
-         * aggregate of them */
-        if (NULL == trk->info) {
-            trk->info = info;
-            trk->ninfo = ninfo;
-        } else {
-            niptr = trk->ninfo + ninfo;
-            PMIX_INFO_CREATE(iptr, niptr);
-            for (n=0; n < trk->ninfo; n++) {
-                PMIX_INFO_XFER(&iptr[n], &trk->info[n]);
-            }
-            for (n=0; n < ninfo; n++) {
-                PMIX_INFO_XFER(&iptr[n+trk->ninfo], &info[n]);
-            }
-            PMIX_INFO_FREE(trk->info, trk->ninfo);
-            trk->info = iptr;
-            trk->ninfo = niptr;
-            /* cleanup */
-            PMIX_INFO_FREE(info, ninfo);
-            info = NULL;
-        }
-        /* add any grpinfo that might have been included */
-        if (NULL != g) {
-            pmix_list_append(&trk->grpinfo, &g->super);
-        }
-    }
-
-    /* add this contributor to the tracker so they get
-     * notified when we are done */
-    pmix_list_append(&trk->local_cbs, &cd->super);
-
-    /* are we locally complete? */
-    if (trk->def_complete && pmix_list_get_size(&trk->local_cbs) == trk->nlocal) {
-        locally_complete = true;
-    }
-
-    /* if we are not locally complete AND this operation
-     * is completely local AND someone specified a timeout,
-     * then we will monitor the timeout in this library.
-     * Otherwise, any timeout must be done by the host
-     * to avoid a race condition whereby we release the
-     * tracker object while the host is still using it */
-    if (!locally_complete && trk->local &&
-        0 < tv.tv_sec && !trk->event_active) {
-        PMIX_THREADSHIFT_DELAY(trk, grp_timeout, tv.tv_sec);
-        trk->event_active = true;
-    }
-
-    /* if we are not locally complete, then we are done */
-    if (!locally_complete) {
-        return PMIX_SUCCESS;
-    }
-
-    /* if all local contributions have been received,
-     * shutdown the timeout event if active */
-    if (trk->event_active) {
-        pmix_event_del(&trk->ev);
-    }
-
-    /* let the local host's server know that we are at the
-     * "fence" point - they will callback once the barrier
-     * across all participants has been completed */
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "local group op complete with %d procs",
-                        (int) trk->npcs);
-
-    if (trk->local) {
-        /* we have created the local group, so we technically
-         * are done. However, we want to give the host a chance
-         * to know about the group to support further operations.
-         * For example, a tool might want to query the host to get
-         * the IDs of existing groups. So if the host supports
-         * group operations, pass this one up to it but indicate
-         * it is strictly local */
-        if (NULL != pmix_host_server.group) {
-            /* we only need to pass the group ID, members, and
-             * an info indicating that this is strictly a local
-             * operation */
-            if (!force_local) {
-                /* add the local op flag to the info array */
-                ninfo = trk->ninfo + 1;
-                PMIX_INFO_CREATE(info, ninfo);
-                for (n=0; n < trk->ninfo; n++) {
-                    PMIX_INFO_XFER(&info[n], &trk->info[n]);
-                }
-                PMIX_INFO_LOAD(&info[trk->ninfo], PMIX_GROUP_LOCAL_ONLY, NULL, PMIX_BOOL);
-                PMIX_INFO_FREE(trk->info, trk->ninfo);
-                trk->info = info;
-                trk->ninfo = ninfo;
-                info = NULL;
-                ninfo = 0;
-            }
-            rc = pmix_host_server.group(PMIX_GROUP_CONSTRUCT, grpid, trk->pcs, trk->npcs,
-                                        trk->info, trk->ninfo, grpcbfunc, trk);
-            if (PMIX_SUCCESS != rc) {
-                if (PMIX_OPERATION_SUCCEEDED == rc) {
-                    /* let the grpcbfunc threadshift the result */
-                    grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-                    return PMIX_SUCCESS;
-                }
-                /* remove the tracker from the list */
-                pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-                PMIX_RELEASE(trk);
-                return rc;
-            }
-            /* we will take care of the rest of the process when the
-             * host returns our call */
-            return PMIX_SUCCESS;
-        } else {
-            /* let the grpcbfunc threadshift the result */
-            grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-            return PMIX_SUCCESS;
-        }
-    }
-
-    /* we don't have to worry about the timeout event being
-     * active in the rest of this code because we only come
-     * here if the operation is NOT completely local, and
-     * we only activate the timeout if it IS local */
-
-    /* check if our host supports group operations */
-    if (NULL == pmix_host_server.group) {
-        /* cannot support it */
-        pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-        PMIX_RELEASE(trk);
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    /* if they direct us to not embed a barrier, then we won't gather
-     * the data for distribution */
-    if (!barrier_directive_included ||
-        (barrier_directive_included && embed_barrier) ||
-        0 < pmix_list_get_size(&trk->grpinfo)) {
-        /* collect any remote contributions provided by group members */
-        PMIX_CONSTRUCT(&bucket, pmix_buffer_t);
-        rc = _collect_data(trk, &bucket);
-        if (PMIX_SUCCESS != rc) {
-            /* remove the tracker from the list */
-            pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-            PMIX_RELEASE(trk);
-            PMIX_DESTRUCT(&bucket);
-            return rc;
-        }
-        /* xfer the results to a byte object */
-        PMIX_UNLOAD_BUFFER(&bucket, bo.bytes, bo.size);
-        PMIX_DESTRUCT(&bucket);
-        /* load any results into a data object for inclusion in the
-         * fence operation */
-        if (0 < bo.size ||
-            0 < pmix_list_get_size(&trk->grpinfo)) {
-            n2 = trk->ninfo + 1; // include space for endpt data
-            PMIX_INFO_CREATE(iptr, n2);
-            for (n = 0; n < trk->ninfo; n++) {
-                PMIX_INFO_XFER(&iptr[n], &trk->info[n]);
-            }
-            /* add the endpt data */
-            PMIX_CONSTRUCT(&bucket, pmix_buffer_t);
-            if (0 < bo.size) {
-                p = pmix_hash_lookup_key(UINT32_MAX, PMIX_GROUP_ENDPT_DATA, NULL);
-                PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &p->index, 1, PMIX_UINT32);
-                if (PMIX_SUCCESS != rc) {
-                    PMIX_ERROR_LOG(rc);
-                    PMIX_DESTRUCT(&bucket);
-                    goto error;
-                }
-                PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &bo, 1, PMIX_BYTE_OBJECT);
-                if (PMIX_SUCCESS != rc) {
-                    PMIX_ERROR_LOG(rc);
-                    PMIX_DESTRUCT(&bucket);
-                    goto error;
-                }
-                PMIX_BYTE_OBJECT_DESTRUCT(&bo);
-            }
-            if (0 < pmix_list_get_size(&trk->grpinfo)) {
-                p = pmix_hash_lookup_key(UINT32_MAX, PMIX_GROUP_INFO, NULL);
-                PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &p->index, 1, PMIX_UINT32);
-                if (PMIX_SUCCESS != rc) {
-                    PMIX_ERROR_LOG(rc);
-                    PMIX_DESTRUCT(&bucket);
-                    goto error;
-                }
-                PMIX_CONSTRUCT(&bkt, pmix_buffer_t);
-                PMIX_LIST_FOREACH(g, &trk->grpinfo, pmix_grpinfo_t) {
-                    PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bkt, &g->blob, 1, PMIX_BYTE_OBJECT);
-                    if (PMIX_SUCCESS != rc) {
-                        PMIX_ERROR_LOG(rc);
-                        PMIX_DESTRUCT(&bucket);
-                        PMIX_DESTRUCT(&bkt);
-                        goto error;
-                    }
-                }
-                PMIX_UNLOAD_BUFFER(&bkt, bo.bytes, bo.size);
-                PMIX_BFROPS_PACK(rc, pmix_globals.mypeer, &bucket, &bo, 1, PMIX_BYTE_OBJECT);
-                if (PMIX_SUCCESS != rc) {
-                    PMIX_ERROR_LOG(rc);
-                    PMIX_DESTRUCT(&bucket);
-                    goto error;
-                }
-                PMIX_BYTE_OBJECT_DESTRUCT(&bo);
-            }
-            PMIX_UNLOAD_BUFFER(&bucket, bo.bytes, bo.size);
-            PMIX_INFO_LOAD(&iptr[n2-1], PMIX_GROUP_ENDPT_DATA, &bo, PMIX_BYTE_OBJECT);
-            PMIX_BYTE_OBJECT_DESTRUCT(&bo);
-            /* replace the tracker's info array */
-            PMIX_INFO_FREE(trk->info, trk->ninfo);
-            trk->info = iptr;
-            trk->ninfo = n2;
-            iptr = NULL;
-        }
-    }
-    rc = pmix_host_server.group(PMIX_GROUP_CONSTRUCT, grpid, trk->pcs, trk->npcs,
-                                trk->info, trk->ninfo, grpcbfunc, trk);
-    if (PMIX_SUCCESS != rc) {
-        if (PMIX_OPERATION_SUCCEEDED == rc) {
-            /* let the grpcbfunc threadshift the result */
-            grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-            return PMIX_SUCCESS;
-        }
-        /* remove the tracker from the list */
-        pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-        PMIX_RELEASE(trk);
-        return rc;
-    }
-
-    return PMIX_SUCCESS;
-
-error:
-    if (NULL != info) {
-        PMIX_INFO_FREE(info, ninfo);
-    }
-    if (NULL != iptr) {
-        PMIX_INFO_FREE(iptr, n2);
-    }
-    return rc;
-}
-
-/* we are being called from the PMIx server's switchyard function,
- * which means we are in an event and can access global data */
-pmix_status_t pmix_server_grpdestruct(pmix_server_caddy_t *cd, pmix_buffer_t *buf)
-{
-    pmix_peer_t *peer = (pmix_peer_t *) cd->peer;
-    int32_t cnt, m;
-    pmix_status_t rc;
-    char *grpid = NULL;
-    pmix_info_t *info = NULL, *iptr;
-    size_t n, ninfo, ninf, niptr;
-    pmix_server_trkr_t *trk;
-    pmix_proc_t *members = NULL;
-    size_t nmembers = 0;
-    bool force_local = false;
-    bool match;
-    bool locally_complete = false;
-    pmix_peer_t *pr;
-    struct timeval tv = {0, 0};
-
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "recvd grpdestruct cmd");
-
-    /* unpack the group ID */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &grpid, &cnt, PMIX_STRING);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-
-    /* is this a failed group? */
-    if (NULL != pmix_server_globals.failedgrps) {
-        for (m=0; NULL != pmix_server_globals.failedgrps[m]; m++) {
-            if (0 == strcmp(grpid, pmix_server_globals.failedgrps[m])) {
-                /* yes - reject it */
-                free(grpid);
-                return PMIX_ERR_TIMEOUT;
-            }
-        }
-    }
-
-    /* unpack the number of members */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &nmembers, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    if (0 == nmembers) {
-        /* not allowed */
-        rc = PMIX_ERR_BAD_PARAM;
-        goto error;
-    }
-    /* unpack the membership */
-    PMIX_PROC_CREATE(members, nmembers);
-    cnt = nmembers;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, members, &cnt, PMIX_PROC);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_PROC_FREE(members, nmembers);
-        goto error;
-    }
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ninf, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    ninfo = ninf + 1;
-    PMIX_INFO_CREATE(info, ninfo);
-    /* store default response */
-    rc = PMIX_SUCCESS;
-    PMIX_INFO_LOAD(&info[ninf], PMIX_LOCAL_COLLECTIVE_STATUS, &rc, PMIX_STATUS);
-    if (0 < ninf) {
-        cnt = ninf;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto error;
-        }
-    }
-
-    /* check directives */
-    for (n = 0; n < ninfo; n++) {
-        if (PMIX_CHECK_KEY(&info[n], PMIX_GROUP_LOCAL_ONLY)) {
-            force_local = PMIX_INFO_TRUE(&info[n]);
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_GROUP_LOCAL_ONLY)) {
-            force_local = PMIX_INFO_TRUE(&info[n]);
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_TIMEOUT)) {
-            PMIX_VALUE_GET_NUMBER(rc, &info[n].value, tv.tv_sec, uint32_t);
-            if (PMIX_SUCCESS != rc) {
-                return rc;
-            }
-        }
-    }
-
-    /* find/create the local tracker for this operation */
-    trk = get_tracker(grpid, members, nmembers, PMIX_GROUP_DESTRUCT_CMD);
-    if (NULL == trk) {
-        /* If no tracker was found - create and initialize it once */
-        trk = new_tracker(grpid, members, nmembers, PMIX_GROUP_DESTRUCT_CMD);
-        if (NULL == trk) {
-            /* only if a bozo error occurs */
-            PMIX_ERROR_LOG(PMIX_ERROR);
-            rc = PMIX_ERROR;
-            goto error;
-        }
-        trk->collect_type = PMIX_COLLECT_NO;
-        /* mark as being a destruct operation */
-        trk->hybrid = true;
-        /* see if this destructor only references local processes */
-        trk->local = true;
-        for (n = 0; n < nmembers; n++) {
-            /* if this entry references the local procs, then
-             * we can skip it */
-            if (PMIX_RANK_LOCAL_PEERS == members[n].rank ||
-                PMIX_RANK_LOCAL_NODE == members[n].rank) {
-                continue;
-            }
-            /* see if it references a specific local proc */
-            match = false;
-            for (m = 0; m < pmix_server_globals.clients.size; m++) {
-                pr = (pmix_peer_t *) pmix_pointer_array_get_item(&pmix_server_globals.clients, m);
-                if (NULL == pr) {
-                    continue;
-                }
-                if (PMIX_CHECK_NAMES(&members[n], &pr->info->pname)) {
-                    match = true;
-                    break;
-                }
-            }
-            if (!match) {
-                /* this requires a non_local operation */
-                trk->local = false;
-                break;
-            }
-        }
-    }
-
-    /* it is possible that different participants will
-     * provide different attributes, so collect the
-     * aggregate of them */
-    if (NULL == trk->info) {
-        trk->info = info;
-        trk->ninfo = ninfo;
-    } else {
-        niptr = trk->ninfo + ninfo;
-        PMIX_INFO_CREATE(iptr, niptr);
-        for (n=0; n < trk->ninfo; n++) {
-            PMIX_INFO_XFER(&iptr[n], &trk->info[n]);
-        }
-        for (n=0; n < ninfo; n++) {
-            PMIX_INFO_XFER(&iptr[n+trk->ninfo], &info[n]);
-        }
-        PMIX_INFO_FREE(trk->info, trk->ninfo);
-        trk->info = iptr;
-        trk->ninfo = niptr;
-        /* cleanup */
-        PMIX_INFO_FREE(info, ninfo);
-        info = NULL;
-    }
-
-    /* add this contributor to the tracker so they get
-     * notified when we are done */
-    pmix_list_append(&trk->local_cbs, &cd->super);
-
-    /* are we locally complete? */
-    if (trk->def_complete && pmix_list_get_size(&trk->local_cbs) == trk->nlocal) {
-        locally_complete = true;
-    }
-
-    /* if we are not locally complete AND this operation
-     * is completely local AND someone specified a timeout,
-     * then we will monitor the timeout in this library.
-     * Otherwise, any timeout must be done by the host
-     * to avoid a race condition whereby we release the
-     * tracker object while the host is still using it */
-    if (!locally_complete && trk->local &&
-        0 < tv.tv_sec && !trk->event_active) {
-        PMIX_THREADSHIFT_DELAY(trk, grp_timeout, tv.tv_sec);
-        trk->event_active = true;
-    }
-
-    /* if we are not locally complete, then we are done */
-    if (!locally_complete) {
-        return PMIX_SUCCESS;
-    }
-
-    /* if all local contributions have been received,
-     * shutdown the timeout event if active */
-    if (trk->event_active) {
-        pmix_event_del(&trk->ev);
-    }
-
-    /* let the local host's server know that we are at the
-     * "fence" point - they will callback once the barrier
-     * across all participants has been completed */
-    pmix_output_verbose(2, pmix_server_globals.group_output,
-                        "local group destruct complete %d",
-                        (int) trk->nlocal);
-    if (trk->local) {
-        /* we have removed the local group, so we technically
-         * are done. However, we want to give the host a chance
-         * to know remove the group to support further operations.
-         * For example, a tool might want to query the host to get
-         * the IDs of existing groups. So if the host supports
-         * group operations, pass this one up to it but indicate
-         * it is strictly local */
-        if (NULL != pmix_host_server.group) {
-            /* we only need to pass the group ID, members, and
-             * an info indicating that this is strictly a local
-             * operation */
-            if (!force_local) {
-                /* add the local op flag to the info array */
-                ninfo = trk->ninfo + 1;
-                PMIX_INFO_CREATE(info, ninfo);
-                for (n=0; n < trk->ninfo; n++) {
-                    PMIX_INFO_XFER(&info[n], &trk->info[n]);
-                }
-                PMIX_INFO_LOAD(&info[trk->ninfo], PMIX_GROUP_LOCAL_ONLY, NULL, PMIX_BOOL);
-                PMIX_INFO_FREE(trk->info, trk->ninfo);
-                trk->info = info;
-                trk->ninfo = ninfo;
-                info = NULL;
-                ninfo = 0;
-            }
-            rc = pmix_host_server.group(PMIX_GROUP_DESTRUCT, grpid,
-                                        members, nmembers,
-                                        trk->info, trk->ninfo, grpcbfunc, trk);
-            if (PMIX_SUCCESS != rc) {
-                if (PMIX_OPERATION_SUCCEEDED == rc) {
-                    /* let the grpcbfunc threadshift the result */
-                    grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-                    PMIX_PROC_FREE(members, nmembers);
-                    free(grpid);
-                    return PMIX_SUCCESS;
-                }
-                /* remove the tracker from the list */
-                pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-                PMIX_RELEASE(trk);
-                PMIX_PROC_FREE(members, nmembers);
-                free(grpid);
-                return rc;
-            }
-            /* we will take care of the rest of the process when the
-             * host returns our call */
-        } else {
-            /* let the grpcbfunc threadshift the result and remove
-             * the group from our list */
-            grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-            PMIX_PROC_FREE(members, nmembers);
-            free(grpid);
-            return PMIX_SUCCESS;
-        }
-    }
-
-    /* this operation requires global support, so check if our host
-     * supports group operations */
-    if (NULL == pmix_host_server.group) {
-        /* cannot support it */
-        pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-        PMIX_RELEASE(trk);
-        PMIX_PROC_FREE(members, nmembers);
-        free(grpid);
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    rc = pmix_host_server.group(PMIX_GROUP_DESTRUCT, grpid,
-                                members, nmembers,
-                                trk->info, trk->ninfo, grpcbfunc, trk);
-    if (PMIX_SUCCESS != rc) {
-        if (PMIX_OPERATION_SUCCEEDED == rc) {
-            /* let the grpcbfunc threadshift the result */
-            grpcbfunc(PMIX_SUCCESS, NULL, 0, trk, NULL, NULL);
-            PMIX_PROC_FREE(members, nmembers);
-            free(grpid);
-            return PMIX_SUCCESS;
-        }
-        /* remove the tracker from the list */
-        pmix_list_remove_item(&pmix_server_globals.collectives, &trk->super);
-        PMIX_RELEASE(trk);
-        PMIX_PROC_FREE(members, nmembers);
-        free(grpid);
-        return rc;
-    }
-
-    PMIX_PROC_FREE(members, nmembers);
-    free(grpid);
-    return PMIX_SUCCESS;
-
-error:
-    if (NULL != info) {
-        PMIX_INFO_FREE(info, ninfo);
-    }
-    if (NULL != members) {
-        PMIX_PROC_FREE(members, nmembers);
-    }
-    if (NULL != grpid) {
-        free(grpid);
-    }
-    return rc;
-}
-
 static void _fabric_response(int sd, short args, void *cbdata)
 {
     pmix_query_caddy_t *qcd = (pmix_query_caddy_t *) cbdata;
@@ -5395,6 +4304,7 @@ static void tcon(pmix_server_trkr_t *t)
     t->collect_type = PMIX_COLLECT_INVALID;
     t->modexcbfunc = NULL;
     t->op_cbfunc = NULL;
+    t->info_cbfunc = NULL;
     t->hybrid = false;
     t->cbdata = NULL;
 }
